+++ drake/bindings/pydrake/autodiffutils.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import, division, print_function
 
-from ._pydrake_autodiffutils import *
+from ._autodiffutils_py import *
