+++ drake/bindings/pydrake/common.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from ._pydrake_common import *
+from ._common_py import *
 
 # When running from python, turn DRAKE_ASSERT and DRAKE_DEMAND failures into
 # SystemExit, instead of process aborts.  See RobotLocomotion/drake#5268.
