+++ drake/bindings/pydrake/mathematicalprogram_py.cc
@@ -5,7 +5,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include "drake/bindings/pydrake/pydrake_symbolic_types.h"
+#include "drake/bindings/pydrake/symbolic_types_py.h"
 #include "drake/solvers/mathematical_program.h"
 #include "drake/solvers/solver_type_converter.h"
 
@@ -78,7 +78,7 @@ auto RegisterBinding(py::handle* pscope,
 
 }  // namespace
 
-PYBIND11_MODULE(_pydrake_mathematicalprogram, m) {
+PYBIND11_MODULE(_mathematicalprogram_py, m) {
   m.doc() = "Drake MathematicalProgram Bindings";
 
   py::object variable =
