+++ drake/bindings/pydrake/mosek_py.cc
