+++ drake/bindings/pydrake/solvers/ik.py
@@ -2,23 +2,23 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 
-from ._pydrake_ik import (IKResults,
-                          IKoptions,
-                          InverseKin,
-                          InverseKinTraj,
-                          InverseKinPointwise,
-                          PostureConstraint,
-                          RelativePositionConstraint,
-                          WorldEulerConstraint,
-                          WorldQuatConstraint,
-                          WorldGazeDirConstraint,
-                          WorldGazeTargetConstraint,
-                          RelativeGazeDirConstraint,
-                          MinDistanceConstraint,
-                          QuasiStaticConstraint)
+from ._ik_py import (IKResults,
+                     IKoptions,
+                     InverseKin,
+                     InverseKinTraj,
+                     InverseKinPointwise,
+                     PostureConstraint,
+                     RelativePositionConstraint,
+                     WorldEulerConstraint,
+                     WorldQuatConstraint,
+                     WorldGazeDirConstraint,
+                     WorldGazeTargetConstraint,
+                     RelativeGazeDirConstraint,
+                     MinDistanceConstraint,
+                     QuasiStaticConstraint)
 
-from ._pydrake_ik import WorldPositionConstraint as _WorldPositionConstraint
-from ._pydrake_ik import WorldPositionInFrameConstraint as \
+from ._ik_py import WorldPositionConstraint as _WorldPositionConstraint
+from ._ik_py import WorldPositionInFrameConstraint as \
     _WorldPositionInFrameConstraint
