+++ drake/bindings/pydrake/symbolic_types_py.h
