+++ drake/bindings/pydrake/test/atlas_constructor_test.py
