+++ drake/bindings/pydrake/test/common_install_test.py
