+++ drake/bindings/pydrake/test/common_test.py
