+++ drake/bindings/pydrake/test/gurobi_solver_test.py
