+++ drake/bindings/pydrake/test/ipopt_solver_test.py
