+++ drake/bindings/pydrake/test/mathematicalprogram_test.py
