+++ drake/bindings/pydrake/test/mosek_solver_test.py
