+++ drake/bindings/pydrake/test/package_map_test.py
