+++ drake/bindings/pydrake/test/rbt_com_test.py
