+++ drake/bindings/pydrake/test/rbt_transform_points_test.py
