+++ drake/bindings/pydrake/test/symbolic_test.py
